remove some types
