+++ apps/meteor/app/livechat/server/lib/localTypes.ts
@@ -1,4 +1,4 @@
-import type { IOmnichannelRoom, IUser, ILivechatVisitor } from '@rocket.chat/core-typings';
+import type { IOmnichannelRoom, IUser, ILivechatVisitor, IMessage, MessageAttachment, IMessageInbox } from '@rocket.chat/core-typings';
 
 export type GenericCloseRoomParams = {
 	room: IOmnichannelRoom;
@@ -29,3 +29,27 @@ export type CloseRoomParamsByVisitor = {
 } & GenericCloseRoomParams;
 
 export type CloseRoomParams = CloseRoomParamsByUser | CloseRoomParamsByVisitor;
+
+type UploadedFile = {
+	_id: string;
+	name?: string;
+	type?: string;
+	size?: number;
+	description?: string;
+	identify?: { size: { width: number; height: number } };
+	format?: string;
+};
+
+export interface ILivechatMessage {
+	token: string;
+	_id: string;
+	rid: string;
+	msg: string;
+	file?: UploadedFile;
+	files?: UploadedFile[];
+	attachments?: MessageAttachment[];
+	alias?: string;
+	groupable?: boolean;
+	blocks?: IMessage['blocks'];
+	email?: IMessageInbox['email'];
+}
